requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
tushare==1.4.3
pymysql==1.1.0
pandas
numpy==1.26.2
requests==2.31.0
matplotlib
mplfinance==0.12.10b0
openai==1.3.7
cryptography==41.0.7
